+++ ChangeLog.md
@@ -7,11 +7,11 @@
   * Send all log output through a new Shell object that formats and adds timestamps.
   * Load ey.yml to customize deploy settings.
   * Control maintenance page through ey.yml options:
-  * * maintenance\_on\_restart: true or false (default: false except for glassfish and mongrel)
-  * * maintenance\_on\_migrate: true or false (default: true)
+    * maintenance\_on\_restart: true or false (default: false except for glassfish and mongrel)
+    * maintenance\_on\_migrate: true or false (default: true)
   * Don't remove maintenance pages that weren't put up during this deploy if maintenance options (above) are not enabled.
   * Control asset precompilation through ey.yml option:
-  * * precompile\_assets: true or false (default: nil - means assets compilation will be inferred by app/assets and config/application.rb)
+    * precompile\_assets: true or false (default: nil - means assets compilation will be inferred by app/assets and config/application.rb)
   * Specify ignore\_database\_adapter\_warning: true to ignore the missing database adapter warning.
 
 ## v1.6.4 (2012-04-26)